refactor + add 'FileBacked' class
