+++ src/gwt/src/org/rstudio/studio/client/RStudioGinjector.java
@@ -18,10 +18,11 @@ import com.google.gwt.core.client.GWT;
 import com.google.gwt.inject.client.GinModules;
 import com.google.gwt.inject.client.Ginjector;
 
-import org.rstudio.core.client.command.AceCommandManager;
+import org.rstudio.core.client.command.EditorCommandManager;
 import org.rstudio.core.client.command.ShortcutManager;
 import org.rstudio.core.client.command.ShortcutViewer;
 import org.rstudio.core.client.command.UserCommandManager;
+import org.rstudio.core.client.files.FileBacked;
 import org.rstudio.core.client.files.filedialog.PathBreadcrumbWidget;
 import org.rstudio.core.client.widget.CaptionWithHelp;
 import org.rstudio.core.client.widget.LocalRepositoriesWidget;
@@ -143,7 +144,8 @@ public interface RStudioGinjector extends Ginjector
    void injectMembers(ModifyKeyboardShortcutsWidget widget);
    void injectMembers(ShortcutManager manager);
    void injectMembers(UserCommandManager manager);
-   void injectMembers(AceCommandManager manager);
+   void injectMembers(EditorCommandManager manager);
+   void injectMembers(FileBacked<?> object);
    
    public static final RStudioGinjector INSTANCE = GWT.create(RStudioGinjector.class);
